requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
transformers==4.35.2
torch==2.1.1
pandas==2.1.3
requests==2.31.0
yfinance==0.2.31
plotly==5.18.0
streamlit==1.32.0
numpy==1.26.2
scikit-learn==1.3.2
ta==0.11.0
newsapi-python==0.2.7
psycopg2-binary==2.9.9
alembic==1.12.1 